Refactor: Update Android Manifest for improved theme and launch behavior
Updated the Android Manifest to improve the theme application and launch behavior. Specifically, set the theme to `NormalTheme` for both the activity and splash screen. Additionally removed package from manifest. Added `taskAffinity` in `MainActivity`. The change removed some of the flutter meta data and replaced it with a single `NormalTheme` metadata.